fix: Fix the expanding toolbar in ArticleDetailActivity
Created a duplicate of the ArticleDetailActivity called
ArticleViewActivity in order to rewrite the activity
logic gradually, without breaking the existing code.
The new activity’s layout is based on the CoordinatorLayout.

The ArticleViewActivity is using support library components
for the up button and the floating button. That provides
better backward compatibility and reduces the app code.