fix: Improve the article text browsing performance
Including a RecyclerView into a NestedScrollView has
caused creation of all items in the RecyclerView. That
has caused several performance issues. Fixed that by
replacing the NestedScrollView that holds the article
title and text with a RecyclerView. The first item of
the RecyclerView holds the title.

Extracted parsing the Article text and splitting to
paragraphs to an AsyncTask and separated paragraphs
into its own LiveData object. This way, the parsing
and the text update in the RecyclerView could be done
in a background task. That also contributes to a better
app performance and makes the UI more responsive.