fix: Improve image loading
The starter code is using a custom image loader implementation.
The article details page is using Picasso library. For some
reason, Picasso did not load the images in some cases or load
them after opening the article again. This is an experiment
to use Glide image loading library. It seems to be working
better. It also requires less code comparing to the current
code for image loading.