fix: Add the image transition
This change is based on the following article:
https://android-developers.googleblog.com/2018/02/continuous-shared-element-transitions.html
Except, the article is using one activity and two fragments.
In this project, there are two activities.

In this change, a transition from the grid to the article is
working. However, the return transition needs more work.

The transition start up time is a bit too slow
probably due to loading data in the article view activity.
This needs to be optimized.