fix: Make UI more responsive
In the current app version, the transition between activities
takes too long. That delays transitions. That might be improved
by using ViewModel for the list of articles and updating the
article list refresh logic. Currently, the articles are refreshed
on a navigation back to ArticleListActivity. That breaks the
return transition. In order to fix the issue, this change
provides some prep work: an upgrade to AndroidX, and an initial
version of ArticleListViewModel with a test for it. The refactoring
of the ArticleListActivity to use the ArticleListViewModel will
be done in the following change.