fix: Fix ArticleViewActivity return transition
Replaced a CursorLoader with a ViewModel in ArticleListActivity.
That prevents reloading data and renders UI faster.

Added SwipeRefreshLayout listener and moved refresh of the data
on the swipe only.

Implemented an initial data loading in ActicleListViewModel. This
way, all updates to the list of articles will be handled
asynchronously. The ArticleListActivity will be notified about
the updates via the LiveData observer.

The navigation back from ArticleViewActivity to ArticleListActivity
is much more smoother now.